Solving tickets #2013 and #2019. Refactoring OutcomeMilestones A & V.
Now it saves the incoming changes for the milestones, according to the
status of the milestone chosen. If the status is "Complete", the reason
now is set as null. If the status is "Extended", the "extendedYear" is
saved, else is set as null. If the status is other than "Complete" and
the reason is "Other", the "otherReason" is saved, else is set as null.

Refactoring Action y Validator of OutcomeMilestones.