Clearer way to show an extended milestone:
On the top left corner, where it shows the unique id and a milestone
"name", now when a milestone has an extended year it says "Extended to"
and the extended year. If the target year and the new target year are
the same (!), the milestone is not considered to be extended and it will
not show the "Extended to" text.