Changes on Table 2 (Policy), 3 (OICR), 4 (Innovation):
- When excluding an OICR, all the linked policies and innovations will
be checked: if the policy/innovation has more studies linked to it and
not a single one of them are included on the AR document, the
policy/innovation will be excluded from the document too.
- If the innovation/policy has no study included on the AR Document, the
user will not be able to include it, as the check box will be not
editable.