Change regarding the "No clear lead" checkbox:
- When "No clear lead" is selected, the field "Lead organization" when
saved will be set to null if it has any institution linked to it. The
validator will also not require this field to be filled. The field "Top
5 organizations" will be required to be filled and will require the user
to select at least 2 and a maximum of 5 institutions.
- When "No clear lead" is NOT selected, the field "Lead" organization
will be now required to the user to be filled. The field "Top 5
organizations" will now not be required and will accept maximum 5
institutions.