Deliverable WOS synchronization save - V2, Part 7
- Added new URL parameter: deliverableID
- Saving info after WOS synchronization: The affiliations will be saved
according to the confidence percentage established by the new variable
created on APConstants. If an affiliation has a confidence level less
than the acceptation percentage, it will be saved on
deliverable_affiliations_not_mapped table; if its confidence level is
greater or equal to the acceptation percentage, it will be saved on
deliverable_affiliation.